initial python-side changes for working in python3 appengine runtime
