+++ index.html
@@ -22,7 +22,7 @@
 <!-- The Google JavaScript loader tool (for the Maps, Charts APIs). -->
   <script src="https://www.google.com/jsapi"></script>
 
-  <script src="/static/ee_api_js.js"></script>
+  <script src="/static/ee_api_js_debug.js"></script>
   <!-- Accessibility Tool javascript. -->
   <script src="/static/access-tool-client-main.js"></script>
